Create easy1
The code determines the length of the last word in a given string. It first calculates the total length of the string using the 'strlen' function and then initializes two variables, 'i' and 'j', to point to the end of the string.
It then iterates backwards from the end of the string, skipping trailing spaces, until it encounters a non-space character. This marks the end of the last word.
Next, it continues iterating backwards until it encounters a space, marking the beginning of the last word.
Finally, it calculates the length of the last word by subtracting the index of the space from the index of the non-space character.